remove all xmtp-node-sdk usage
