+++ package.json
@@ -8,7 +8,6 @@
 		"dev": "bun --watch server.ts",
 		"start": "bun clean-installations && bun wipe-db && bun server.ts",
 		"gen-keys": "bun scripts/gen-keys",
-		"check-installations": "bun scripts/check-installations",
 		"clean-installations": "bun scripts/clean-installations",
 		"wipe-db": "bun scripts/wipe-db",
 		"check": "biome check",
@@ -24,7 +23,6 @@
 		"@xmtp/content-type-text": "^2.0.2",
 		"@xmtp/content-type-transaction-reference": "^2.0.2",
 		"@xmtp/content-type-wallet-send-calls": "^2.0.0",
-		"@xmtp/node-sdk": "^4.0.3",
 		"ai": "4.3.19",
 		"dotenv": "^17.2.0",
 		"uint8arrays": "^5.1.0",
